refactor: split up engine into smaller manageable files
