+++ src/main/modules/journey/engine/navigation.ts
@@ -1,0 +1,88 @@
+import { JourneyConfig, StepConfig } from './schema';
+
+export class Navigation {
+  /**
+   * Simplified next step resolution
+   */
+  static resolveNext(step: StepConfig, allData: Record<string, unknown>): string {
+    const nxt = step.next;
+    if (!nxt) {
+      return step.id;
+    }
+    if (typeof nxt === 'string') {
+      return nxt;
+    }
+
+    const stepData = allData[step.id] as Record<string, unknown>;
+    if (!stepData) {
+      return nxt.else || step.id;
+    }
+
+    // If when is a function, evaluate it directly
+    if (typeof nxt.when === 'function') {
+      try {
+        const conditionMet = (nxt.when as (sd: Record<string, unknown>, ad: Record<string, unknown>) => boolean)(
+          stepData ?? {},
+          allData ?? {}
+        );
+        return conditionMet ? nxt.goto : nxt.else || step.id;
+      } catch {
+        // Logging should be done by caller if needed
+        return nxt.else || step.id;
+      }
+    }
+
+    // If we reach here and when is not a function (or threw), default to else or stay
+    return nxt.else || step.id;
+  }
+
+  /**
+   * Find the previous step for back navigation by analyzing journey flow
+   */
+  static findPreviousStep(
+    currentStepId: string,
+    journey: JourneyConfig,
+    allData: Record<string, unknown>
+  ): string | null {
+    // Find which step(s) can lead to the current step
+    for (const [stepId, stepConfig] of Object.entries(journey.steps)) {
+      const typedStepConfig = stepConfig as StepConfig;
+      const next = typedStepConfig.next;
+
+      if (!next) {
+        continue;
+      }
+
+      // Direct next step
+      if (typeof next === 'string' && next === currentStepId) {
+        return stepId;
+      }
+
+      // Conditional next step
+      if (typeof next === 'object') {
+        const stepData = allData[stepId] as Record<string, unknown>;
+        if (!stepData) {
+          continue;
+        }
+
+        // Handle functional conditions first
+        if (typeof next.when === 'function') {
+          const conditionMet = (next.when as (sd: Record<string, unknown>, ad: Record<string, unknown>) => boolean)(
+            stepData ?? {},
+            allData ?? {}
+          );
+          if (conditionMet && next.goto === currentStepId) {
+            return stepId;
+          }
+          if (!conditionMet && next.else === currentStepId) {
+            return stepId;
+          }
+        }
+
+        // If next.when is not a function we cannot reliably determine previous; skip
+      }
+    }
+
+    return null;
+  }
+}
